refactor(ramp): require paths
