+++ app/components/UI/Ramp/components/ApplePayButton.tsx
@@ -6,8 +6,8 @@ import Text from '../../../Base/Text';
 import StyledButton from '../../StyledButton';
 
 /* eslint-disable import/no-commonjs, @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports */
-const ApplePayLogoLight = require('../../../../../images/ApplePayLogo-light.png');
-const ApplePayLogoDark = require('../../../../../images/ApplePayLogo-dark.png');
+const ApplePayLogoLight = require('../../../../images/ApplePayLogo-light.png');
+const ApplePayLogoDark = require('../../../../images/ApplePayLogo-dark.png');
 /* eslint-enable import/no-commonjs, @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports */
 
 const styles = StyleSheet.create({
